requirements.txt
fastapi
uvicorn
opencv-python-headless
insightface
onnxruntime
rapidocr_onnxruntime
faiss-cpu
numpy
python-multipart
